feat (core): create base button
